User now display only their brackets (based on the groups they are in)
